Restrict access using GitHub authentication
This will only allow access to users in a certain GitHub organisation.
The GitHub application and organisation to use are configurable via
environment variables.

The `warden-github-rails` gem stores the authenticated GitHub user's
details in the session. Rails 4.1 changes the default session store to
be JSON-based, which isn't capable of storing complex objects. I've
switched back to the old Marshal-based session store to make it work.