Use rails_12factor gem when running on Heroku
Heroku provide a gem which modifies your app's configuration for better
compatibility with their platform. It sets up the Rails logger to write
to STDOUT, and enables some Rack middleware to serve static assets.

I'd like these things to work when running on Heroku, but I also don't
want to make it harder to run Banjaxed on other platforms. Here the gem
will only be required if it looks we're running on Heroku, based on the
environment variables present.

This reverts 5948ec47b08c8ee4826829870f039f7f90629f97, which I did last
week to make assets work, in favour of letting the gem take care of it.